Moved EnumExtensions to the EnumExtensions folder
